Create Minimum Pairwise Product Sum.cpp
Antonio and his girlfriend Teena decided to move their relationship to the next level and decided to buy arrays from a nearby arrays shop.

Each of them bought one array. Antonio bought array A consisting of N integers and Teena bought array B consisting of M integers.

Now Antonio and Teena are compatible with each other(that’s why they are in a relationship) and want their arrays to have maximum compatibility too.

Compatibility of two arrays A and B is measured as :

Two arrays are compatible if and only if they have an equal number of elements.
Compatibility value of arrays A and B each having length L = A1*B1 + A2*B2 + … + AL*BL.
To achieve this Antonio and Teena decided to remove some elements from each of their arrays such that arrays A and B have an equal number of elements and their compatibility value is as large as possible.
Sample Input 1

sizes of both arrays = 3, 4

arr1= -1, 2, 1

arr2=-100, 3, 10, -9

Sample Output 1

120

Explanation of Sample 1

Erase the 3rd element of array A. Array A becomes = [-1, 2]

Erase the 2nd and 4th element of array A. Array A becomes = [-100, 10]

Compatibility values of these arrays = [(-1)*(-100) + 2*10] = 120.